webgui: Move to new package fs/rc/webgui.
